+++ albomon/core/src/Infrastructure/UI/Cli/Command/CheckAlboPopCatalogComand.php
@@ -38,7 +38,7 @@ class CheckAlboPopCatalogComand extends Command
     protected function execute(InputInterface $input, OutputInterface $output): ?int
     {
         $alboList = $this->getCustomCatalog();
-        // TODO remove
+
         $AlboPopSpecValidation = 'Non Rilevato';
 
         $io = $this->getSymfonyStyle($input, $output);
@@ -68,7 +68,6 @@ class CheckAlboPopCatalogComand extends Command
                 if (!$monitorResult->httpStatus()) {
                     $table->appendRow([$monitorResult->feedUrl(), sprintf('<error>%s</error>', 'NON ATTIVO'), $AlboPopSpecValidation, '', 'server error']);
                 } else {
-                    // TODO format output per data aggiornamento visualizzare colore diverso per ritardo maggiore di 1 settimana  1 mese 1 anno
                     $lastFeedItemDateWithDifference = $this->formatContentUpdatedAt($monitorResult->lastFeedItemDate());
 
                     $table->appendRow([$monitorResult->feedUrl(), 'ATTIVO', $AlboPopSpecValidation, $lastFeedItemDateWithDifference, '']);